csp-apps/tc/tc/requirements_devel.txt
pytest>=3.3.2
pytest-django>=3.1.2
